Improve preview generation, matrix mentions and image support
This change redesigns the interface of plugins which are now expected to
provide two methods:

- format_content

  which takes the parsed markdown content and returns a
  (formatted_content, preview, warnings) tuple, in which
  - formatted_content is a plugin-specific object,
  - preview is the plugin's best string representation of what a post generated
    by it would look like
  - warnings is a string of any plugin-specific warning messages that were
    produced during processing of the content

- create_post

  which takes the formatted_content generated by format_content and
submits it as a post to the respective platform;
  returns a success, link tuple, where success can be true or false, and
link is a link to the generated post on the platform or None if no link
is available

Matrix mentions are now formatted much more according to the matrix
specs (including the mentioned user's display name if it can be found)
and will now display correctly in at least two clients (element and
nheko). Likewise, images are now uploaded in a format compatible with
both of these clients.