Rewritten the bluesky and slack plugins and add some fixes.
This commit updates the dependencies in the `requirements.txt` file. It also updates the GitHub Actions workflows (`preview.yml` and `publish_content.yml`) to use the latest versions of the `actions/checkout` and `actions/setup-python` actions. The Python version in the workflows is updated to `3.12` (this was necessary for  using `asyncio.Runner()`) . These changes aim to improve the overall functionality and compatibility of the project.